admission: move kvSlotAdjuster into its own file
Pure code movement. We're hurting some git history tracking with these
code changes, but hopefully they lead to more understandability with
future work.

Release note: None
